move metrics stream initialization into seperate file
